+++ components/point/api/src/commonMain/kotlin/ru/kyamshanov/mission/components/points/api/di/TaskComponent.kt
@@ -1,0 +1,9 @@
+package ru.kyamshanov.mission.components.points.api.di
+
+import ru.kyamshanov.mission.components.points.api.domain.SearchTaskUseCase
+import ru.kyamshanov.mission.components.points.api.presentation.navigation.TaskLauncher
+
+interface TaskComponent {
+    val searchTaskUseCase: SearchTaskUseCase
+    val launcher: TaskLauncher
+}
