+++ components/point/api/src/commonMain/kotlin/ru/kyamshanov/mission/components/points/api/domain/SearchTaskUseCase.kt
@@ -1,0 +1,9 @@
+package ru.kyamshanov.mission.components.points.api.domain
+
+import ru.kyamshanov.mission.components.points.api.common.TaskSlim
+
+
+interface SearchTaskUseCase {
+
+    suspend fun getAll(): Result<List<TaskSlim>>
+}
