--- components/point/api/src/commonMain/kotlin/ru/kyamshanov/mission/components/points/api/editing/di/EditingTaskComponent.kt
@@ -1,5 +1,0 @@
-package ru.kyamshanov.mission.components.points.api.editing.di
-
-interface EditingTaskComponent {
-    val launcher: EditingTaskLauncher
-}
