--- components/point/api/src/commonMain/kotlin/ru/kyamshanov/mission/components/points/api/editing/di/EditingTaskLauncher.kt
@@ -1,8 +1,0 @@
-package ru.kyamshanov.mission.components.points.api.editing.di
-
-import ru.kyamshanov.mission.components.points.api.common.TaskId
-
-interface EditingTaskLauncher {
-
-    fun launch(taskId: TaskId)
-}
