+++ components/point/api/src/commonMain/kotlin/ru/kyamshanov/mission/components/points/api/presentation/navigation/TaskLauncher.kt
@@ -1,0 +1,10 @@
+package ru.kyamshanov.mission.components.points.api.presentation.navigation
+
+import ru.kyamshanov.mission.components.points.api.common.TaskId
+
+interface TaskLauncher {
+
+    fun launchEditing(taskId: TaskId)
+
+    fun launchCreationTask()
+}
