--- components/point/api/src/commonMain/kotlin/ru/kyamshanov/mission/components/points/api/search/di/SearchTaskComponent.kt
@@ -1,8 +1,0 @@
-package ru.kyamshanov.mission.components.points.api.search.di
-
-import ru.kyamshanov.mission.components.points.api.search.domain.usecases.SearchTaskUseCase
-
-interface SearchTaskComponent {
-
-    val searchTaskUseCase: SearchTaskUseCase
-}
