--- components/point/api/src/commonMain/kotlin/ru/kyamshanov/mission/components/points/api/search/domain/usecases/SearchTaskUseCase.kt
@@ -1,9 +1,0 @@
-package ru.kyamshanov.mission.components.points.api.search.domain.usecases
-
-import ru.kyamshanov.mission.components.points.api.common.TaskSlim
-
-
-interface SearchTaskUseCase {
-
-    suspend fun getAll(): Result<List<TaskSlim>>
-}
