+++ components/point/impl/src/commonMain/kotlin/ru/kyamshanov/mission/components/point/impl/data/model/TaskPriorityDto.kt
@@ -1,0 +1,8 @@
+package ru.kyamshanov.mission.components.point.impl.data.model
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+enum class TaskPriorityDto {
+    PRIMARY
+}
