+++ components/point/impl/src/commonMain/kotlin/ru/kyamshanov/mission/components/point/impl/data/model/TaskStatusDto.kt
@@ -1,0 +1,9 @@
+package ru.kyamshanov.mission.components.point.impl.data.model
+
+import kotlinx.serialization.Serializable
+
+
+@Serializable
+enum class TaskStatusDto {
+    CREATED, IN_PROCESSING, COMPLETED
+}
