--- components/point/impl/src/commonMain/kotlin/ru/kyamshanov/mission/components/point/impl/editing/data/model/TaskStatusDto.kt
@@ -1,9 +1,0 @@
-package ru.kyamshanov.mission.components.point.impl.editing.data.model
-
-import kotlinx.serialization.Serializable
-
-
-@Serializable
-enum class TaskStatusDto {
-    CREATED, IN_PROCESSING, COMPLETED
-}
