--- components/point/impl/src/commonMain/kotlin/ru/kyamshanov/mission/components/point/impl/editing/presentation/EditingTaskScreen.kt
@@ -1,6 +1,0 @@
-package ru.kyamshanov.mission.components.point.impl.editing.presentation
-
-import ru.kyamshanov.mission.components.points.api.common.TaskId
-import ru.kyamshanov.mission.core.navigation.api.Screen
-
-internal expect class EditingTaskScreen actual constructor(taskId: TaskId) : Screen
