+++ components/point/impl/src/commonMain/kotlin/ru/kyamshanov/mission/components/point/impl/presentation/EditingTaskScreen.kt
@@ -1,0 +1,6 @@
+package ru.kyamshanov.mission.components.point.impl.presentation
+
+import ru.kyamshanov.mission.components.points.api.common.TaskId
+import ru.kyamshanov.mission.core.navigation.api.Screen
+
+internal expect class EditingTaskScreen actual constructor(taskId: TaskId) : Screen
