+++ components/point/impl/src/commonMain/kotlin/ru/kyamshanov/mission/components/point/impl/presentation/TaskLauncherImpl.kt
@@ -1,0 +1,17 @@
+package ru.kyamshanov.mission.components.point.impl.presentation
+
+import ru.kyamshanov.mission.components.points.api.common.TaskId
+import ru.kyamshanov.mission.components.points.api.presentation.navigation.TaskLauncher
+import ru.kyamshanov.mission.core.navigation.api.Navigator
+
+internal class TaskLauncherImpl(
+    private val navigator: Navigator
+) : TaskLauncher {
+    override fun launchEditing(taskId: TaskId) {
+        navigator.navigateTo(EditingTaskScreen(taskId))
+    }
+
+    override fun launchCreationTask() {
+        TODO("Not yet implemented")
+    }
+}
