--- components/point/impl/src/commonMain/kotlin/ru/kyamshanov/mission/components/point/impl/search/data/api/PointApi.kt
@@ -1,8 +1,0 @@
-package ru.kyamshanov.mission.components.point.impl.search.data.api
-
-import ru.kyamshanov.mission.components.point.impl.search.data.model.AttachedTasksResponseDto
-
-internal interface PointApi {
-
-    suspend fun allProjects(): AttachedTasksResponseDto
-}
