--- components/point/impl/src/commonMain/kotlin/ru/kyamshanov/mission/components/point/impl/search/data/model/TaskPriorityDto.kt
@@ -1,8 +1,0 @@
-package ru.kyamshanov.mission.components.point.impl.search.data.model
-
-import kotlinx.serialization.Serializable
-
-@Serializable
-enum class TaskPriorityDto {
-    PRIMARY
-}
